Almost finish DetailVC and fix item model

+++ FoodAppMVC/Model/Item.swift
@@ -11,6 +11,6 @@ struct Item {
     var name: String?
     var imageName: String?
     var description: String?
-    var price: Int?
+    var price: Double?
     var weight: Int?
 }
